+++ CHANGELOG.md
@@ -97,6 +97,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Allow jsonnet imports in pipeline configuration
 
 ## [2.0.4]
 ### Fixed